+++ streams/block/ciphers/des/single_des.h
@@ -28,7 +28,7 @@ namespace block {
             : block_cipher(rounds)
             , _ctx(encrypt) { }
 
-        void keysetup(const std::uint8_t* key, const std::uint64_t keysize) override;
+        void keysetup(const std::uint8_t* key, const std::size_t keysize) override;
 
         void ivsetup(const std::uint8_t* iv, const uint64_t ivsize) override;
