+++ streams/block/ciphers/des/triple_des.h
@@ -28,9 +28,9 @@ namespace block {
             : block_cipher(rounds)
             , _ctx(encrypt) {}
 
-        void keysetup(const std::uint8_t* key, const std::uint64_t keysize) override;
+        void keysetup(const std::uint8_t* key, const std::size_t keysize) override;
 
-        void ivsetup(const std::uint8_t* iv, const std::uint64_t ivsize) override;
+        void ivsetup(const std::uint8_t* iv, const std::size_t ivsize) override;
 
         void encrypt(const std::uint8_t* plaintext,
                      std::uint8_t* ciphertext) override;
