+++ streams/block/ciphers/tea/tea.h
@@ -24,9 +24,9 @@ namespace block {
         tea(std::size_t rounds)
             : block_cipher(rounds) {}
 
-        void keysetup(const std::uint8_t* key, const std::uint64_t keysize) override;
+        void keysetup(const std::uint8_t* key, const std::size_t keysize) override;
 
-        void ivsetup(const std::uint8_t* iv, const std::uint64_t ivsize) override;
+        void ivsetup(const std::uint8_t* iv, const std::size_t ivsize) override;
 
         void encrypt(const std::uint8_t* plaintext,
                      std::uint8_t* ciphertext) override;
